add reset and revise linker.ld
