+++ debug.gdb
@@ -1,0 +1,2 @@
+file build/kernel8.elf
+target remote :1234
